Solution to Calculate Average Marks of Students

+++ Calculate Average Marks of Students@@67877d61dc107486ad5c2743/code.c
@@ -5,16 +5,16 @@
 typedef struct student {
     int rollnumber;
     char name[100];
-    int marks;
+    float marks;
 } student;
 
 int main(){
     int a;
     scanf("%d", &a);
     student s[a];
-    int summ = 0;
+    float summ = 0;
     for(int i =0; i<a; i++){
         summ += s[i].marks;
     }
-    printf("Average Marks: %d", summ);
+    printf("Average Marks: %.2f", summ);
 }
